Have task-related types under a sub-package
