+++ plugin-base/src/main/java/io/spine/tools/gradle/task/DynamicTaskName.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.gradle;
+package io.spine.tools.gradle.task;
 
 import io.spine.value.StringTypeValue;
