+++ plugin-base/src/main/java/io/spine/tools/gradle/task/ProtobufTaskName.java
@@ -23,38 +23,29 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.tools.gradle;
 
-import io.spine.annotation.Internal;
+package io.spine.tools.gradle.task;
 
 /**
- * Names of Gradle tasks defined by the {@code base} plugin.
+ * Names of Gradle tasks defined by the Protobuf Gradle plugin.
  *
- * @see <a href="https://docs.gradle.org/current/userguide/base_plugin.html#sec:base_tasks">
- *         the plugin doc</a>
+ * @see <a href="https://github.com/google/protobuf-gradle-plugin">the plugin doc</a>
  */
-@Internal
-public enum BaseTaskName implements TaskName {
+public enum ProtobufTaskName implements TaskName {
 
     /**
-     * Deletes the temporary build artifacts.
+     * Generates production code from Protobuf.
+     *
+     * <p>Note that this task is not a public API of the plugin. Users should be conscious and
+     * cautious when depending on it.
      */
-    clean,
+    generateProto,
 
     /**
-     * Aggregate task that assembles all the artifacts of this project.
+     * Generates test code from Protobuf.
+     *
+     * <p>Note that this task is not a public API of the plugin. Users should be conscious and
+     * cautious when depending on it.
      */
-    assemble,
-
-    /**
-     * A lifecycle task which marks the project verification routines, such as static code analysis,
-     * executing tests, etc.
-     */
-    check,
-
-    /**
-     * A lifecycle task which builds everything in the project, including running tests, producing
-     * production artifacts, and generating documentation.
-     */
-    build
+    generateTestProto
 }
