+++ plugin-base/src/main/java/io/spine/tools/gradle/task/TaskDependencies.java
@@ -23,7 +23,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.tools.gradle;
+package io.spine.tools.gradle.task;
 
 import org.gradle.api.Task;
