+++ plugin-base/src/main/java/io/spine/tools/gradle/task/package-info.java
@@ -24,32 +24,13 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.gradle;
-
-import io.spine.annotation.Internal;
-
 /**
- * A name of a Gradle task.
+ * This package provides types and utilities for working with Gradle tasks.
  */
-@Internal
-public interface TaskName {
+@CheckReturnValue
+@ParametersAreNonnullByDefault
+package io.spine.tools.gradle.task;
 
-    /**
-     * The value of the name.
-     *
-     * <p>If an enum implements this interface, it is expected to name its constants so that
-     * the {@link Enum#name()} obtains the name of the task.
-     */
-    String name();
+import com.google.errorprone.annotations.CheckReturnValue;
 
-    /**
-     * Obtains this task name as a path.
-     *
-     * <p>It is expected that the referred task belongs to the root project (a.k.a {@code :}).
-     *
-     * @return the name with a colon symbol ({@code :}) at the beginning
-     */
-    default String path() {
-        return ':' + name();
-    }
-}
+import javax.annotation.ParametersAreNonnullByDefault;
