+++ plugin-testlib/src/main/java/io/spine/tools/gradle/testing/TaskArguments.java
@@ -29,7 +29,7 @@ package io.spine.tools.gradle.testing;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
-import io.spine.tools.gradle.TaskName;
+import io.spine.tools.gradle.task.TaskName;
 
 import java.util.List;
